requirements.txt
asgiref==3.4.1
attrs==21.2.0
cffi==1.14.6
coverage==5.5
cryptography==3.4.8
Django==3.2.7
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.12.1
hypothesis==6.20.0
inflection==0.5.1
iniconfig==1.1.1
jwt==1.2.0
Markdown==3.3.4
mixer==7.1.2
packaging==21.0
pluggy==1.0.0
py==1.10.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
pytest-factoryboy==2.1.0
pytest-sugar==0.9.4
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
